+++ packages/charts/react-charts/library/src/types/DataPoint.ts
@@ -210,6 +210,12 @@ export interface VerticalBarChartDataPoint {
    * Accessibility data for callout
    */
   callOutAccessibilityData?: AccessibilityProps;
+
+  /**
+   * Label to display on the bar
+   * This is an optional prop that can be used to show custom text on bars
+   */
+  barLabel?: string;
 }
 
 /**
@@ -636,6 +642,12 @@ export interface VSChartDataPoint {
    * The prop used to define the culture to localized the numbers
    */
   culture?: string;
+
+  /**
+   * Label to display on the bar
+   * This is an optional prop that can be used to show custom text on bars
+   */
+  barLabel?: string;
 }
 
 /**
@@ -746,6 +758,11 @@ export interface GVBarChartSeriesPoint {
    * False by default.
    */
   useSecondaryYScale?: boolean;
+
+  /**
+   * Label to display on the bar
+   */
+  barLabel?: string;
 }
 
 /**