+++ packages/react-components/react-menu-grid-preview/library/CHANGELOG.md
@@ -1,8 +1,21 @@
 # Change Log - @fluentui/react-menu-grid-preview
 
-This log was last generated on Fri, 05 Dec 2025 22:38:36 GMT and should not be manually modified.
+This log was last generated on Wed, 17 Dec 2025 18:10:10 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [0.2.7](https://github.com/microsoft/fluentui/tree/@fluentui/react-menu-grid-preview_v0.2.7)
+
+Wed, 17 Dec 2025 18:10:10 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-menu-grid-preview_v0.2.6..@fluentui/react-menu-grid-preview_v0.2.7)
+
+### Patches
+
+- Bump @fluentui/react-menu to v9.20.6 ([PR #35582](https://github.com/microsoft/fluentui/pull/35582) by beachball)
+- Bump @fluentui/react-table to v9.19.6 ([PR #35582](https://github.com/microsoft/fluentui/pull/35582) by beachball)
+- Bump @fluentui/react-tabster to v9.26.11 ([PR #35582](https://github.com/microsoft/fluentui/pull/35582) by beachball)
+- Bump @fluentui/react-jsx-runtime to v9.3.4 ([PR #35582](https://github.com/microsoft/fluentui/pull/35582) by beachball)
+- Bump @fluentui/react-utilities to v9.26.0 ([PR #35582](https://github.com/microsoft/fluentui/pull/35582) by beachball)
 
 ## [0.2.6](https://github.com/microsoft/fluentui/tree/@fluentui/react-menu-grid-preview_v0.2.6)